requirements.txt
pandas
numpy
librosa
Scikit-learn
streamlit
matplotlib
pillow
audio-recorder-streamlit
wave